ShadowFramework: sharing version tagged as 2.0.2.201. You will find some documented changes in ShadowFramework2.0_OpenGL20/docs/version2.0.2.201.pdf.  SF2.0Light and SF20LiteTestWorld has been aligned with the new version. SF20LiteTestWorld also keeps a lot of test/demos/esamples of the new functionalities.

+++ ShadowFrameworkLite20/src/shadow/math/SFVertex3f.java
@@ -87,7 +87,7 @@ public class SFVertex3f extends SFVertex2f{
 		v[2]=vx.v[2];
 	}
 	
-	public double dot3f(SFVertex3f vx){
+	public float dot3f(SFVertex3f vx){
 		return vx.v[0]*v[0]+vx.v[1]*v[1]+vx.v[2]*v[2];
 	}
